backend/core/requirements.txt
asgiref==3.5.1
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
pymongo==4.1.1
python-snappy==0.6.1
pytz==2022.1
sqlparse==0.2.4
typing_extensions==4.2.0
winkerberos==0.9.0
